requirements.txt
# This file lists the external Python package dependencies required by the
# ComfyUI-Holaf custom nodes. These packages must be installed in the
# Python environment where ComfyUI is running for all nodes to function correctly.
#
# To install all dependencies at once, run the following command in your
# ComfyUI's Python environment/venv:
#
# pip install -r requirements.txt
#
# Alternatively, you can install them individually:
# pip install numpy Pillow spandrel pandas matplotlib psutil
#
# Note:
# - This list excludes standard Python libraries (e.g., os, math, json).
# - This list excludes packages assumed to be installed by ComfyUI itself
#   (e.g., torch, torchvision) or its core dependencies.

numpy
Pillow
spandrel
pandas
matplotlib
psutil